feat: replace Firestore bonusProducts with GET /bonus/v2/products
