+++ src/modules/bonus/queries/index.tsx
@@ -5,3 +5,4 @@ export {
   getBonusAmountEarnedQueryKey,
 } from './use-bonus-amount-earned-query';
 export {useProductPointsQuery} from './use-product-points-query';
+export {useActiveBonusProductsQuery} from './use-active-bonus-products-query';
