+++ src/modules/bonus/use-relevant-bonus-product.ts
@@ -1,0 +1,15 @@
+import {FormFactor} from '@atb/api/types/generated/mobility-types_v2';
+import {useActiveBonusProductsQuery} from './queries';
+
+export const useRelevantBonusProduct = (
+  operatorId: string | undefined,
+  formFactor: FormFactor,
+) => {
+  const {data: activeBonusProducts} = useActiveBonusProductsQuery();
+  if (!operatorId) return undefined;
+  return activeBonusProducts?.find(
+    (bonusProduct) =>
+      bonusProduct.formFactors.includes(formFactor) &&
+      bonusProduct.operatorIds.includes(operatorId),
+  );
+};
