+++ src/modules/mobility/components/sheets/BicycleSheet.tsx
@@ -24,11 +24,10 @@ import {
 import {TransportationIconBox} from '@atb/components/icon-box';
 import {useAnalyticsContext} from '@atb/modules/analytics';
 import {
-  findRelevantBonusProduct,
   PayWithBonusPointsCheckbox,
   useIsBonusActiveForUser,
+  useRelevantBonusProduct,
 } from '@atb/modules/bonus';
-import {useFirestoreConfigurationContext} from '@atb/modules/configuration';
 import {Vehicle} from '@atb/api/types/mobility';
 import {VehicleCard} from '../VehicleCard';
 import {PriceDetailsCard} from '../PriceDetailsCard';
@@ -92,12 +91,7 @@ export const BicycleSheet = ({
   const {operatorBenefit} = useOperatorBenefit(operatorId);
   const [payWithBonusPoints, setPayWithBonusPoints] = useState(false);
   const {logEvent} = useAnalyticsContext();
-  const {bonusProducts} = useFirestoreConfigurationContext();
-  const bonusProduct = findRelevantBonusProduct(
-    bonusProducts,
-    operatorId,
-    FormFactor.Bicycle,
-  );
+  const bonusProduct = useRelevantBonusProduct(operatorId, FormFactor.Bicycle);
   const {mode, subMode} = getTransportModeAndSubMode(
     FormFactor.Bicycle,
     vehicle?.vehicleType.propulsionType,
@@ -229,7 +223,7 @@ export const BicycleSheet = ({
                   />
                 </View>
               )}
-              {!!isBonusActiveForUser && bonusProduct && (
+              {isBonusActiveForUser && !!bonusProduct && (
                 <PayWithBonusPointsCheckbox
                   bonusProduct={bonusProduct}
                   operatorName={operatorName}
